+++ src/services/email.js
@@ -23,4 +23,14 @@ const emailService = {
   },
 };
 
-module.exports = { emailService };
+const fakeEmailService = {
+  async sendVerificationEmail({ toEmail, verificationCode }) {
+    console.log(`Mock: Sending verification email to ${toEmail}. Code: ${verificationCode}`);
+    return true;
+  },
+};
+
+module.exports = {
+  emailService,
+  fakeEmailService,
+};
