--- src/services/fakeEmail.js
@@ -1,8 +1,0 @@
-const fakeEmailService = {
-  async sendVerificationEmail({ toEmail, verificationCode }) {
-    console.log(`Mock: Sending verification email to ${toEmail}. Code: ${verificationCode}`);
-    return true;
-  },
-};
-
-module.exports = { fakeEmailService };
